+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Execution of Ansible playbook that starts platform services.
     - Waiting for kubernetes pods to reach expected states.
     - Boot of the compute nodes and User Access Nodes (UANs) using BOS.
+- Debug logging for every stage of `sat bootsys` operations including duration
+  of each stage.
 
 ### Fixed
 - Fixed missing values in info messages logged from sat.filtering module.